Add files via upload
Stores people's names  in a stack. Asks repeatedly the user integer numbers from 1 to 4, and according to the user-informed number, pops this amount of people, printing the name of the one remaining in the top of the stack. For example, if the user enters 2, removes two people from the stack and print the "third" remaining in the stack The program stops as the stack becomes empty.